requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
bilibili-dynamic==0.0.8
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
frozenlist==1.3.3
idna==2.10
kiwisolver==1.4.4
matplotlib==3.6.2
multidict==6.0.2
numpy==1.23.5
packaging==21.3
pathlib==1.0.1
Pillow==8.4.0
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
qrcode==6.1
requests==2.23.0
six==1.16.0
typing_extensions==4.4.0
urllib3==1.25.11
yarl==1.8.1
